added examples for darcall and freyja
