+++ example/mixture_composition/config_freyja.yaml
@@ -1,0 +1,66 @@
+sequencing: paired_end
+samples:                                # List of sample categories to be analysed
+    - name: .*                          # Regex expression of sample names to be analysed (reads/original/example.*_R1.fastq.gz)
+      reference: sars_cov_2             # Reference genome for reads in the category (reference/mhv/mhv.fa)
+
+report_dir: report/public/01-mixture_freyja # Generated reports and essential output files would be stored there
+threads: 16                             # Number of threads to use in analysis
+
+reads:                                  # Prepare reads and quality reports for downstream analysis
+    preprocess:                         # Pre-process of reads, eliminate sequencing artifacts, contamination ...
+
+        trimmed:                        # Remove low quality parts of reads
+            method: cutadapt
+            temporary: True
+            fasta: reference/sars_cov_2/adapters/artic_v4.fa
+            minimum_length: 45          # shorter reads are discarded
+            quality_cutoff: 20,20       # cutoff on 3' end ("20") or cutoff on both 5' and 3' ends ("20,20")
+            length: 500                 # truncate reads to this length
+            head_cut: 15                # number of bases that are removed from the beginning of each read
+            tail_cut: 0                 # number of bases are removed from the end of each read
+
+        decontaminated:                 # Eliminate fragments from known artificial source, e.g. contamination by human
+            method: bowtie2             # Supported values: bowtie2
+            temporary: True             # If True, generated files would be removed after successful analysis
+            references:                 # List of reference genomes
+                - grch38_decoy
+            keep: False                 # Keep reads mapped to references (True) or remove them as contamination (False)
+
+    report:                             # Summary reports of read characteristics to assess their quality
+        quality_report:                 # HTML summary report of read quality
+            method: fastqc              # Supported values: fastqc
+            read_types:                 # List of preprocess steps for quality reports
+                - original
+                - trimmed
+                - decontaminated
+
+mapping:                                # Find the most similar genomic region to reads in reference (mapping process)
+    mapper:                             # Method for mapping
+        method: bwa                     # Supported values: bowtie2, bwa
+        temporary: True
+
+    index:                              # Generate .bai index for mapped reads in .bam files
+        method: samtools                # Supported values: samtools
+
+    postprocess:                        # Successive steps to refine mapped reads
+        sorted:
+            method: samtools
+            temporary: True
+
+        deduplicated:                   # Mark duplicated reads (PCR duplicated)
+            method: picard              # Supported values: picard
+
+    report:                             # Summary reports of mapping process and results
+        quality_report:                 # HTML summary with quality of mappings
+            method: qualimap            # Supported values: qualimap
+            map_types:                  # List of post-process steps for quality reports
+               - deduplicated
+
+mixture_composition:                    # obtain composition of variants in percentage
+    caller:
+        method: freyja                  # will support more methods in the future
+        # updating_dir: ~/freyja_dir/   # in case of updating freyja, set this directory where 'freyja update' outputs the updated data
+        # sif_path: example.sif         # in case of running freyja in singularity, define path to the sif file
+        # singularity_bind:             # in case of running freyja in singularity, define the binding path
+report:
+    multiqc: True
