+++ example/mixture_composition/reference/sars_cov_2/sars_cov_2.fa.fai
@@ -1,0 +1,1 @@
+NC_045512.2	29903	86	80	81
