+++ rules/shared/mixture_composition/caller/darcall.snake
@@ -1,0 +1,22 @@
+rule mixture_composition_darcall:
+    '''
+    Outputs mixture composition
+    '''
+    input:
+        #vcf  = 'variant/{reference}-{panel}/%s/{sample}.vcf' % pipeline.postprocessed_variant_type
+        vcf  = 'variant/{reference}-{panel}/filtered/{sample}.vcf'
+    output:
+        composition_tsv  = 'consensus/{reference}-{panel}/{sample}.mixture_comp.csv'
+    log:
+        out = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.log',
+        err = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.err'
+    params:
+        script = srcdir("darcall_scripts/darcall.py")
+    shell:
+        '''
+        python {params.script} \
+            {input.vcf} \
+            {output.composition_tsv} \
+            1> {log.out} \
+            2> {log.err};
+        '''
