--- rules/shared/mixture_composition/caller/daria.snake
@@ -1,22 +1,0 @@
-rule mixture_composition_daria:
-    '''
-    Outputs mixture composition
-    '''
-    input:
-        #vcf  = 'variant/{reference}-{panel}/%s/{sample}.vcf' % pipeline.postprocessed_variant_type
-        vcf  = 'variant/{reference}-{panel}/filtered/{sample}.vcf'
-    output:
-        composition_tsv  = 'consensus/{reference}-{panel}/{sample}.mixture_comp.csv'
-    log:
-        out = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.log',
-        err = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.err'
-    params:
-        script = srcdir("daria.py")
-    shell:
-        '''
-        python {params.script} \
-            {input.vcf};
-        
-        cp 'summary.tsv' {output.composition_tsv};
-        rm 'summary.tsv';
-        '''
