+++ rules/shared/mixture_composition/caller/freyja.snake
@@ -1,3 +1,12 @@
+import os
+
+def prepare_arg(base:str, arg:str, file:str) -> str:
+    """Prepare optional argument for freyja"""
+    if base is None:
+        return ''
+    path = os.path.join(base, file)
+    return f'{arg} {path}'
+
 rule samtools_mpileup_depth:
     '''
     Builds .depth file
@@ -7,9 +16,6 @@ rule samtools_mpileup_depth:
         fasta = 'reference/{reference}/{reference}.fa',
     output:
         depth   = 'mapping/{reference}-{panel}/freyja/{sample}_freyja.depth'
-    log:
-        out = 'consensus/{reference}-{panel}/log/{sample}.samtools_depth.log',
-        err = 'consensus/{reference}-{panel}/log/{sample}.samtools_depth.err'
     conda:
         config['snakelines_dir'] + '/enviroments/samtools.yaml'
     shell:
@@ -17,9 +23,9 @@ rule samtools_mpileup_depth:
         samtools mpileup -aa -A -d 600000 -Q 20 -q 0 -B -f {input.fasta} {input.bam} | cut -f1-4 > {output.depth}
         '''
 
-rule mixture_composition_freyja:
+rule mixture_composition_freyja_variants:
     '''
-    Outputs mixture composition using freyja
+    Call variants
     '''
     input:
         fasta = 'reference/{reference}/{reference}.fa',
@@ -27,23 +33,74 @@ rule mixture_composition_freyja:
         bam   = 'mapping/{{reference}}/{map_type}/{{sample}}.bam'.format(map_type=pipeline.postprocessed_map_type),
         depth = 'mapping/{reference}-{panel}/freyja/{sample}_freyja.depth',
     output:
+        tsv     = 'consensus/{reference}-{panel}/{sample}_freyja.tsv',
+    log:
+        out = 'consensus/{reference}-{panel}/log/{sample}.freyja_variants.log',
+        err = 'consensus/{reference}-{panel}/log/{sample}.freyja_variants.err'
+    conda:
+        config['snakelines_dir'] + '/enviroments/freyja.yaml'
+    shell:
+        '''
+        freyja variants \
+            {input.bam} \
+            --variants {output.tsv} \
+            --depths {input.depth} \
+            --ref {input.fasta} \
+            1> {log.out} \
+            2> {log.err};
+        '''
+
+rule mixture_composition_freyja_demix:
+    '''
+    Demix variants using freyja method
+    '''
+    input:
+        depth = 'mapping/{reference}-{panel}/freyja/{sample}_freyja.depth',
+        tsv     = 'consensus/{reference}-{panel}/{sample}_freyja.tsv',
+    output:
+        demix  = 'consensus/{reference}-{panel}/{sample}_freyja.demix',
+    log:
+        out = 'consensus/{reference}-{panel}/log/{sample}.freyja_demix.log',
+        err = 'consensus/{reference}-{panel}/log/{sample}.freyja_demix.err'
+    params:
+        lineages = prepare_arg(method_config.get('updating_dir',None),'--meta','curated_lineages.json'),
+        barcodes = prepare_arg(method_config.get('updating_dir',None),'--barcodes','usher_barcodes.csv'),
+    conda:
+        config['snakelines_dir'] + '/enviroments/freyja.yaml'
+    shell:
+        '''
+        freyja demix \
+            {input.tsv} \
+            {input.depth} \
+            --output {output.demix} \
+            {params.lineages} \
+            {params.barcodes} \
+            1> {log.out} \
+            2> {log.err};
+        '''
+
+rule mixture_composition_freyja:
+    '''
+    Postprocess freyja output
+    '''
+    input:
+        demix   = 'consensus/{reference}-{panel}/{sample}_freyja.demix',
+    output:
         composition_tsv  = 'consensus/{reference}-{panel}/{sample}.mixture_comp.csv'
     log:
         out = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.log',
         err = 'consensus/{reference}-{panel}/log/{sample}.mixture_comp.err'
     params:
-        tsv     = 'consensus/{reference}-{panel}/{sample}_freyja.tsv',
-        demix   = 'consensus/{reference}-{panel}/{sample}_freyja.demix',
-        script_tsv = srcdir("convert_demixed_to_tsv.py"),
-        script_variants = srcdir("freyja_variants.sh"),
-        script_demix = srcdir("freyja_demix.sh"),
-        sif_freyja = '/home_pfs/tools/snakelines/repos/latest2_miniconda_freya_image.sif',
-        sample_name = '{sample}',
-        project_dir = '/home_pfs/data/'
+        script_tsv = srcdir("freyja_scripts/convert_demixed_to_tsv.py"),
+    conda:
+        config['snakelines_dir'] + '/enviroments/freyja.yaml'
     shell:
         '''
-        resolved=$(readlink -f "{input.fasta}"); # due to singularity
-        singularity exec --bind {params.project_dir} {params.sif_freyja} bash {params.script_variants} {input.bam} {params.tsv} {input.depth} $resolved;
-        singularity exec --bind {params.project_dir} {params.sif_freyja} bash {params.script_demix} {params.tsv} {input.depth} {params.demix};
-        python3 {params.script_tsv} {params.demix} {params.sample_name} {output.composition_tsv};
+        python3 \
+            {params.script_tsv} \
+            {input.demix} \
+            {wildcards.sample} \
+            {output.composition_tsv} \
+            1> {log.out} \
+            2> {log.err};
         '''
