+++ rules/shared/mixture_composition/caller/freyja_scripts/freyja_demix.sh
@@ -1,0 +1,3 @@
+eval "$(conda shell.bash hook)"
+conda activate freyja-env
+freyja demix $1 $2 --output $3 $4 $5
