+++ rules/shared/mixture_composition/caller/freyja_scripts/freyja_update.sh
@@ -1,0 +1,3 @@
+eval "$(conda shell.bash hook)"
+conda activate freyja-env
+freyja update --outdir $1
