--- rules/shared/mixture_composition/caller/freyja_update.sh
@@ -1,3 +1,0 @@
-eval "$(conda shell.bash hook)"
-conda activate freyja-env
-freyja update --outdir /home_pfs/data/narcos/production/scripts/snakelines/dev/rules/shared/mixture_composition/caller/lineages
