--- rules/shared/mixture_composition/caller/freyja_variants.sh
@@ -1,3 +1,0 @@
-eval "$(conda shell.bash hook)"
-conda activate freyja-env
-freyja variants $1 --variants $2 --depths $3 --ref $4
